moved bank into parallel
