+++ src/mrchem.cpp
@@ -15,7 +15,6 @@
 #include "mrchem.h"
 #include "mrenv.h"
 #include "parallel.h"
-#include "utils/Bank.h"
 
 #include "chemistry/Molecule.h"
 
@@ -48,11 +47,11 @@ int main(int argc, char **argv) {
             for (const auto &json_rsp : json_rsps) driver::run_rsp(json_rsp, mol);
         }
         driver::print_properties(mol);
-        if (mpi::grand_master()) orb_bank.close();
+        if (mpi::grand_master()) mpi::orb_bank.close();
         mpi::barrier(mpi::comm_orb);
         mrenv::finalize(timer.elapsed());
     } else {
-        orb_bank.open();
+        mpi::orb_bank.open();
     }
 
     mpi::finalize();
