+++ src/parallel.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "MRCPP/Parallel"
+#include <map>
 
 #ifdef HAVE_MPI
 #include <mpi.h>
@@ -68,7 +69,60 @@ void allreduce_matrix(ComplexMatrix &mat, MPI_Comm comm);
 
 } // namespace mpi
 
-class Bank;
+struct BankFunctionData {
+    int real_size;
+    int imag_size;
+    bool is_shared;
+};
+
+struct BankOrbitalData {
+    int rank_id;
+    int spin;
+    int occ;
+};
+
+struct deposit {
+    Orbital *orb;
+    BankFunctionData funcinfo; // Function info
+    BankOrbitalData orbinfo;   // Orbital info
+    int id;                    // to identify what is deposited
+    int source;                // mpi rank from the source of the data
+};
+
+struct queue_struct {
+    int id;
+    std::vector<int> clients;
+};
+
+class Bank {
+public:
+    Bank();
+    ~Bank();
+    int open();
+    int close();
+    int clear_all(int i, MPI_Comm comm);
+    int clear(int ix);
+    int put_orb(int id, Orbital &orb);
+    int get_orb(int id, Orbital &orb);
+    int put_func(int id, QMFunction &func);
+    int get_func(int id, QMFunction &func);
+
+private:
+    int const CLOSE_BANK = 1;
+    int const CLEAR_BANK = 2;
+    int const GET_ORBITAL = 3;
+    int const SAVE_ORBITAL = 4;
+    int const GET_FUNCTION = 5;
+    int const SAVE_FUNCTION = 6;
+    int clear_bank();
+    std::map<int, int> id2ix;
+    std::vector<deposit> deposits;
+    std::map<int, int> id2qu;
+    std::vector<queue_struct> queue;
+};
+
+namespace mpi {
 extern Bank orb_bank;
+} // namespace mpi
 
 } // namespace mrchem
