+++ src/qmoperators/one_electron/KineticOperator.cpp
@@ -33,7 +33,8 @@ ComplexMatrix KineticOperator::operator()(OrbitalVector &bra, OrbitalVector &ket
     ComplexMatrix T_tot = ComplexMatrix::Zero(Ni, Nj);
     mrcpp::print::memory(2, "memusage Kinetic start");
 
-    if (false) {
+    // we keep both methods for testing purposes
+    if (mpi::bank_size > 0) {
         Timer timer;
         // 1) save all derivatives
         OrbitalVector dKet = p_x(ket);
@@ -42,60 +43,21 @@ ComplexMatrix KineticOperator::operator()(OrbitalVector &bra, OrbitalVector &ket
         for (int i = 0; i < bra.size(); i++) {
             if (not mpi::my_orb(bra[i])) continue;
             Orbital der = dKet[i];
-            // std::cout<<i<<" size x derivative "<<der.getNNodes(0)<<std::endl;
-            orb_bank.put_orb(i, der);
+            mpi::orb_bank.put_orb(i, der);
         }
-
         mrcpp::print::tree(2, "save xderivatives ", 1, 1, timer.elapsed());
         // 2) symmetric dot product
         mpi::barrier(mpi::comm_orb);
         mrcpp::print::tree(2, "barrier ", 1, 1, timer.elapsed());
-        /*//testing
-        timer.start();
-        if(mpi::orb_rank==0){
-            for (int i = 0; i <  bra.size(); i++) {
-        Orbital der;
-                orb_bank.get_orb(i+Ni, der);
-                std::cout<<i<<" from bank "<< (int)(1000*timer.elapsed())<<" "<<der.getNNodes(0)<<std::endl;
-                timer.start();
-            }
-            int j=0;
-            for (int i = 0; i <  bra.size(); i++) {
-                if(!mpi::my_orb(bra[i]) or j>5) continue;
-                mpi::send_orbital(dKet[i], 4, j);
-                j++;
-                //                std::cout<<j<<" send "<< (int)(1000*timer.elapsed())<<std::endl;
-                timer.start();
-            }
-        }
-        if(mpi::orb_rank==4){
-            int j=0;
-            for (int i = 0; i <  bra.size(); i++) {
-                if(j>5) continue;
-                Orbital der;
-                mpi::recv_orbital(der, 0, j);
-                std::cout<<j<<" recv "<< (int)(1000*timer.elapsed())<<" "<<der.getNNodes(0)<<std::endl;
-                j++;
-                timer.start();
-           }
-        }
-        mpi::barrier(mpi::comm_orb);
-        */
         int shift = mpi::orb_rank;
         for (int ii = shift; ii < bra.size() + shift; ii++) {
             int i = ii % bra.size();
             Orbital deri;
-            if (mpi::orb_rank == 0)
-                std::cout << i << " owner " << bra[i].rankID() << " " << (&bra == &ket) << std::endl;
             if (mpi::my_orb(bra[i])) {
-                if (mpi::orb_rank != bra[i].rankID()) std::cout << i << "ERROR owner " << bra[i].rankID() << std::endl;
                 deri = dKet[i];
             } else {
                 Timer timer;
-                orb_bank.get_orb(i, deri);
-                if (mpi::orb_rank == 0)
-                    std::cout << i << " from bank  " << (int)(1000 * timer.elapsed()) << " " << deri.getNNodes(0)
-                              << std::endl;
+                mpi::orb_bank.get_orb(i, deri);
             }
             //            mrcpp::print::tree(2, "get orbi ", i, 1, timer.elapsed());
             for (int j = 0; j < ket.size(); j++) {
@@ -126,8 +88,7 @@ ComplexMatrix KineticOperator::operator()(OrbitalVector &bra, OrbitalVector &ket
             int i = ii % bra.size();
             if (not mpi::my_orb(bra[i])) continue;
             Orbital der = dKet[i];
-            //            std::cout<<i<<" size y derivative "<<der.getNNodes(0)<<std::endl;
-            orb_bank.put_orb(i, der);
+            mpi::orb_bank.put_orb(i, der);
         }
         mpi::barrier(mpi::comm_orb);
         // 2) symmetric dot product
@@ -136,7 +97,7 @@ ComplexMatrix KineticOperator::operator()(OrbitalVector &bra, OrbitalVector &ket
             if (mpi::my_orb(bra[i])) {
                 deri = dKet[i];
             } else {
-                orb_bank.get_orb(i, deri);
+                mpi::orb_bank.get_orb(i, deri);
             }
             for (int j = 0; j < ket.size(); j++) {
                 if (not mpi::my_orb(ket[j])) continue;
@@ -159,8 +120,7 @@ ComplexMatrix KineticOperator::operator()(OrbitalVector &bra, OrbitalVector &ket
         for (int i = 0; i < bra.size(); i++) {
             if (not mpi::my_orb(bra[i])) continue;
             Orbital der = dKet[i];
-            // std::cout<<i<<" size z derivative "<<der.getNNodes(0)<<std::endl;
-            orb_bank.put_orb(i, der);
+            mpi::orb_bank.put_orb(i, der);
         }
         mpi::barrier(mpi::comm_orb);
         // 2) symmetric dot product
@@ -170,11 +130,10 @@ ComplexMatrix KineticOperator::operator()(OrbitalVector &bra, OrbitalVector &ket
             if (mpi::my_orb(bra[i])) {
                 deri = dKet[i];
             } else {
-                orb_bank.get_orb(i, deri);
+                mpi::orb_bank.get_orb(i, deri);
             }
             for (int j = 0; j < ket.size(); j++) {
                 if (not mpi::my_orb(ket[j])) continue;
-                //                if( (&bra == &ket) and (i>(Ni-1)/2+j or (i>=j-(Ni-1)/2 and i<j))) continue;
                 if ((&bra == &ket) and (i + j + (i > j)) % 2) continue;
                 Orbital derj;
                 if (&bra == &ket) {
@@ -182,84 +141,20 @@ ComplexMatrix KineticOperator::operator()(OrbitalVector &bra, OrbitalVector &ket
                 } else {
                     derj = p_z(ket[j]);
                 }
-                // mrcpp::print::tree(2, "get bankj ", i, j, timer.elapsed());
                 T_z(i, j) = orbital::dot(deri, derj);
-                // mrcpp::print::tree(2, "get dot ", i, j, timer.elapsed());
                 if (&bra == &ket) T_z(j, i) = std::conj(T_z(i, j));
             }
         }
-
-        /*   for (int i = 0; i <  bra.size(); i++) {
-        if (not mpi::my_orb(bra[i])) continue;
-        Orbital der = p_x(bra[i]);
-        std::cout<<i<<" size derivative "<<der.getNNodes(0)<<std::endl;
-        orb_bank.put_orb(i+1*Ni, der);
-        //der.free(NUMBER::Total);
-        der = p_y(bra[i]);
-        orb_bank.put_orb(i+2*Ni, der);
-        //der.free(NUMBER::Total);
-        der = p_z(bra[i]);
-        orb_bank.put_orb(i+3*Ni, der);
-    }
-    mrcpp::print::tree(2, "bank", bra.size(),bra.size(), timer.elapsed());
-    //std::cout<<mpi::orb_rank<<" start  make overlap matrix "<<bra.size()<<" "<<Ni<<std::endl;
-    //3) make overlap matrix
-    for (int i = 0; i <  bra.size(); i++) {
-        Orbital deri;
-        orb_bank.get_orb(i+Ni, deri);
-        mrcpp::print::tree(2, "get bank ", i, bra.size(), timer.elapsed());
-        for (int j = 0; j <  ket.size(); j++) {
-            if (not mpi::my_orb(ket[j])) continue;
-           // std::cout<<mpi::orb_rank<<"  "<<i<<" "<<j<<" i>(Ni-1)/2+j "<<(i>(Ni-1)/2+j)<<"(i>=j-(Ni-1)/2 and i<j)
-    "<<((i>=j-(Ni-1)/2 and i<j))<<std::endl; if( (&bra == &ket) and (i>(Ni-1)/2+j or (i>=j-(Ni-1)/2 and i<j))) continue;
-            Orbital derj;
-            orb_bank.get_orb(j+Nj, derj);
-            mrcpp::print::tree(2, "get bankj ", i, j, timer.elapsed());
-            T_x(i,j) = orbital::dot(deri, derj);
-            mrcpp::print::tree(2, "get dot ", i, j, timer.elapsed());
-            if (&bra == &ket) T_x(j,i) = std::conj(T_x(i,j));
-        }
-        orb_bank.get_orb(i+2*Ni, deri);
-        for (int j = 0; j <  ket.size(); j++) {
-            if (not mpi::my_orb(ket[j])) continue;
-            if( (&bra == &ket) and (i>(Ni-1)/2+j or (i>=j-(Ni-1)/2 and i<j))) continue;
-            Orbital derj;
-            orb_bank.get_orb(j+2*Nj, derj);
-           T_y(i,j) = orbital::dot(deri, derj);
-           if (&bra == &ket) T_y(j,i) = std::conj(T_y(i,j));
-        }
-        orb_bank.get_orb(i+3*Ni, deri);
-        for (int j = 0; j <  ket.size(); j++) {
-            if (not mpi::my_orb(ket[j])) continue;
-            if( (&bra == &ket) and (i>(Ni-1)/2+j or (i>=j-(Ni-1)/2 and i<j))) continue;
-            Orbital derj;
-            orb_bank.get_orb(j+3*Nj, derj);
-            T_z(i,j) = orbital::dot(deri, derj);
-            if (&bra == &ket) T_z(j,i) = std::conj(T_z(i,j));
-        }
-        }*/
         T_tot = 0.5 * (T_x + T_y + T_z);
         mpi::allreduce_matrix(T_tot, mpi::comm_orb);
         mrcpp::print::memory(2, "memusage Kinetic end1");
-        // if(mpi::orb_rank==0)std::cout<<"KINMATRIX bank "<<std::endl<<T_tot<<std::endl;
-        if (mpi::orb_rank == 0)
-            ;
         return T_tot;
     } else {
-        //}{
-        // mpi::allreduce_matrix(T_y, mpi::comm_orb);
-        //    if(mpi::orb_rank==0)std::cout<<"KINMATRIX bank "<<std::endl<<T_tot<<std::endl;
-        // T_x.setZero();
-        // T_y.setZero();
-        // T_z.setZero();
-
         {
             Timer timer;
             int nNodes = 0, sNodes = 0;
             if (&bra == &ket) {
                 OrbitalVector dKet = p_x(ket);
-                std::cout << mpi::orb_rank << " size derivative " << orbital::get_n_nodes(dKet) << " in " << ket.size()
-                          << std::endl;
                 nNodes += orbital::get_n_nodes(dKet);
                 sNodes += orbital::get_size_nodes(dKet);
                 T_x = orbital::calc_overlap_matrix(dKet);
@@ -312,9 +207,6 @@ ComplexMatrix KineticOperator::operator()(OrbitalVector &bra, OrbitalVector &ket
             }
             mrcpp::print::tree(2, "<i|p[z]p[z]|j>", nNodes, sNodes, timer.elapsed());
         }
-
-        // if (mpi::orb_rank == 0) std::cout << "KINMATRIX orig " << std::endl << (0.5 * (T_x + T_y + T_z)) <<
-        // std::endl;
         mrcpp::print::memory(2, "memusage Kinetic end2");
         return 0.5 * (T_x + T_y + T_z);
     }
